Removido y reorganizado algunos fragmentos de código viejo, comentada la funcionalidad de algunos componentes
