+++ src/main/java/com/planeador/controlador/AdministradorControlador.java
@@ -30,6 +30,7 @@ import com.planeador.modelo.Administrador;
 import com.planeador.modelo.Docente;
 import com.planeador.modelo.Materia;
 
+//Este controlador administra todas las rutas del aplicativo que se van como /admin/{lo que sea}
 @Controller
 @RequestMapping("/admin")
 public class AdministradorControlador {
@@ -41,8 +42,9 @@ public class AdministradorControlador {
 	@Autowired
 	private MateriaServicio MateriaServicio;
 	
-//	De manera temporal para rastrear las sesiones de Admin, buscar una
+//	De manera temporal para rastrear el objeto Admin a lo largo de la sesión, buscar una
 //	solución más estable en el futuro
+	
 	@ModelAttribute("admin")
 	public Administrador getAdministradorActual(Model model, HttpServletRequest request) {
 		Administrador admin = (Administrador) model.getAttribute("admin");
@@ -64,85 +66,68 @@ public class AdministradorControlador {
 	}
 
 	@GetMapping("")
-	public String login(HttpServletRequest request, HttpSession session, Model model) {
+	public String rutaPorDefectoAdministrador(HttpServletRequest request, HttpSession session, Model model) {
 		if (request.getSession().getAttribute("admin_id") != null) {
 			return "main";
 		} else
 			return "login_admin";
 	}
 
-//	@GetMapping("login")
-//	public String loginPage(HttpServletRequest request, HttpSession session, Model model) {
-////		request.getSession().setAttribute("es_admin", true);
-//		request.getSession().invalidate();
-//		return "login_admin";
-//	}
-
 	@PostMapping("/login")
-	public String validate(RedirectAttributes att, @RequestParam String email, @RequestParam String password,
+	public String validarInicioSesionAdmin(RedirectAttributes att, @RequestParam String email, @RequestParam String password,
 			HttpServletRequest request, HttpSession session, Model model) {
 
 		Administrador admin = administradorService.select(email, password);
 
 		if (admin != null) {
+			
+			// Tuve que resetear el valor "docente_id" para solucionar un bug extraño con los inicios de sesión, pendiente
+			// revisar a profundidad
+			request.getSession().setAttribute("docente_id", 0);
 			request.getSession().setAttribute("admin_id", admin.getId());
 			model.addAttribute("admin", admin);
-//			En proceso de refactorizar, comentar esta linea si es necesario
-//			request.getSession().setAttribute("admin", admin);
 			return "redirect:/admin/main";
 		} else {
 			att.addFlashAttribute("loginError", "Usuario o contraseña incorrecta");
 			return "redirect:/admin";
 		}
 	}
+	
 
 	@GetMapping("/logout")
-	public String logout(HttpServletRequest request, HttpSession session) {
+	public String cerrarSesionAdmin(HttpServletRequest request, HttpSession session) {
 		request.getSession().invalidate();
 		return "redirect:/admin/";
 	}
 
 	@PostMapping("/save")
-	public String insertAdmin(RedirectAttributes att, Administrador admin) {
+	public String guardarAdmin(RedirectAttributes att, Administrador admin) {
 		administradorService.save(admin);
 		att.addFlashAttribute("accion", "Administrador registrado con éxito!");
 		return "redirect:/admin/";
 	}
 
+//	A desactivar en un futuro cuando no necesitemos crear más administradores.
 	@GetMapping("/new")
-	public String showForm(Model model) {
-		return "registeradmin";
+	public String formularioRegistroAdmin(Model model) {
+		return "register_admin";
 	}
 
-//	Esto está basado en ProductoController.java del ejercicio de refencia, luego
-//	pensamos bien si moverlo a otro lado, básicamente para recuperar los datos
-//	del admin. y mostrarlos en el perfil
-
 	@GetMapping("/main")
-	public String perfilAdministrador(HttpServletRequest request, Model model) {
-//		int admin_id = (int) request.getSession().getAttribute("admin_id");
-//		Administrador adm = this.administradorService.get(admin_id);
-//		NOTA, usar la request garantiza que el objeto esté disponible durante todas las peticiones a la sesión,
-//		pero usar el model solamente lo autoriza para la petición en particular
+	public String seccionPrincipalAdmin(HttpServletRequest request, Model model) {
 		return "main";
 	}
 
-	@GetMapping("/perfil")
-	public String elPerfil(HttpServletRequest request, Model model) {
-//		Testeando el nuevo @ModelAttribute
-//		int admin_id = (int) request.getSession().getAttribute("admin_id");
-//		Administrador adm = this.administradorService.get(admin_id);
-//		model.addAttribute("admin", adm);
-		return "perfil";
-	}
-
 	@GetMapping("/solicitudes")
-	public String solicitudesRegistro(HttpServletRequest request, Model model) {
+	public String listarSolicitudesDocente(HttpServletRequest request, Model model) {
 
 //		https://stackoverflow.com/questions/50348166/java-8-list-of-objects-to-mapstring-list-of-values
 //		https://www.appsloveworld.com/springboot/100/35/how-can-i-create-dynamic-table-in-thymeleaf
+//		El siguiente acomodo intenta generar la tabla de manera dinámica pasándole a la plantilla el nombre de las columnas y usando
+//		un Map para asociarlo con los atributos correspondientes de cada objeto docente. Funciona de momento.
 		List<Docente> solicitudes = this.docenteService.findPendingRequests();
 		List<String> headers = Arrays.asList("Nombre", "Email", "Aprobación");
+		
 		List<Map<String, Object>> rows = solicitudes.stream().map(docente -> {
 			Map<String, Object> um = new HashMap<>();
 			um.put(headers.get(0), docente.getNombre());
@@ -157,7 +142,7 @@ public class AdministradorControlador {
 	}
 
 	@GetMapping("/aprobar/{id}")
-	public String aceptarSolicitud(@PathVariable Integer id, Model model) {
+	public String aprobarSolicitudDocente(@PathVariable Integer id, Model model) {
 		Docente porAprobar = this.docenteService.get(id);
 		porAprobar.setAprobado(true);
 //		Aparentemente este método también actualiza
@@ -168,43 +153,64 @@ public class AdministradorControlador {
 	}
 
 	@GetMapping("/rechazar/{id}")
-	public String rechazarSolicitud(@PathVariable Integer id, Model model) {
+	public String rechazarSolicitudDocente(@PathVariable Integer id, Model model) {
 		this.docenteService.delete(id);
 		model.addAttribute("msjRechazo", "Solicitud aprobada con éxito");
 		return "redirect:/admin/solicitudes/";
 
 	}
 
-//	@GetMapping("/materias")
-//	public String listaDeMaterias (Model model){
-//		List<Materia> materias = this.materiaService.getAll();
-//		 List<String> headers = Arrays.asList("Nombre","# de Créditos","Semestre", "Acción");
-//		// List<Map<String, Object>> rows = materias.stream().map(materia -> {
-//		// 	Map<String, Object> um = new HashMap<>();
-//		// 	um.put(headers.get(0), materia.getNombre());
-//		// 	um.put(headers.get(1), materia.getCreditos());
-//		// 	um.put(headers.get(2), materia.getSemestre());
-//		// 	um.put(headers.get(2), materia.getId());
-//		// 	return um;
-//		// }).collect(Collectors.toList());
-//		 model.addAttribute("headers", headers);
-//		// model.addAttribute("rows", rows);
-//		model.addAttribute("rows", materias);
-//		return "materias";
-//	}
-
 //	https://www.kindsonthegenius.com/part-2-how-to-implement-pagination-in-spring-boot-with-thymeleaf/
 	@GetMapping("/materias")
-	public String nuevaPaginaDeMaterias(
+	public String buscarPaginaDeMaterias(
 			@RequestParam(value = "pagina", required = false, defaultValue = "1") int pagina,
 			@RequestParam(value = "nroDeElementos", required = false, defaultValue = "5") int nroDeElementos,
 			Model model) {
+//		Utiliza un método del servicio para extraer una consulta paginada de la lista de objetos almacenados
+//		en la BBDD. El objeto Page tiene varias propiedades que son de utilidad en la plantilla final
 		Page<Materia> paraElControlador = MateriaServicio.paginaDeMaterias(pagina, nroDeElementos);
-		// Indica el número total de páginas para dibujar en la plantilla
+		// Indica el número total de páginas totales en que se puede dividir toda la lista de objetos traía de la BBDD
 		int totalPages = paraElControlador.getTotalPages();
-		System.out.println(paraElControlador);
 		model.addAttribute("rows", paraElControlador);
 		model.addAttribute("pageNumbers", totalPages);
 		return "materias";
 	}
 }
+
+//EN PAUSA
+//El siguiente código se utilizó en algún punto temprano y se conserva como referencia, pero posiblemente
+//sea desechado en el futuro
+
+//@GetMapping("/perfil")
+//public String elPerfil(HttpServletRequest request, Model model) {
+//	Testeando el nuevo @ModelAttribute
+//	int admin_id = (int) request.getSession().getAttribute("admin_id");
+//	Administrador adm = this.administradorService.get(admin_id);
+//	model.addAttribute("admin", adm);
+//	return "perfil";
+//}
+
+//@GetMapping("/materias")
+//public String listaDeMaterias (Model model){
+//	List<Materia> materias = this.materiaService.getAll();
+//	 List<String> headers = Arrays.asList("Nombre","# de Créditos","Semestre", "Acción");
+//	// List<Map<String, Object>> rows = materias.stream().map(materia -> {
+//	// 	Map<String, Object> um = new HashMap<>();
+//	// 	um.put(headers.get(0), materia.getNombre());
+//	// 	um.put(headers.get(1), materia.getCreditos());
+//	// 	um.put(headers.get(2), materia.getSemestre());
+//	// 	um.put(headers.get(2), materia.getId());
+//	// 	return um;
+//	// }).collect(Collectors.toList());
+//	 model.addAttribute("headers", headers);
+//	// model.addAttribute("rows", rows);
+//	model.addAttribute("rows", materias);
+//	return "materias";
+//}
+
+//@GetMapping("login")
+//public String loginPage(HttpServletRequest request, HttpSession session, Model model) {
+////	request.getSession().setAttribute("es_admin", true);
+//	request.getSession().invalidate();
+//	return "login_admin";
+//}
