+++ src/main/java/com/planeador/modelo/Materia.java
@@ -17,18 +17,16 @@ public class Materia {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
-
+	
 	@Column
 	private String nombre;
-
+	
 	@Column
 	private Integer creditos;
-
+	
 	@Column
 	private Integer semestre;
-
-
-
+	
 	@OneToMany(mappedBy = "materia")
 	List<Microcurriculo> microcurriculos;
