+++ src/main/java/com/planeador/servicio/impl/MateriaServicioImpl.java
@@ -1,4 +1,4 @@
-package com.planeador.servicio;
+package com.planeador.servicio.impl;
 
 import java.util.List;
