+++ src/main/resources/templates/fragments/header.html
@@ -1,11 +1,13 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
+
 <!-- 	https://www.thymeleaf.org/doc/tutorials/3.0/usingthymeleaf.html#flexible-layouts-beyond-mere-fragment-insertion -->
 <!-- https://attacomsian.com/blog/thymeleaf-get-session-attributes -->
 <!-- https://studygyaan.com/spring-boot/fragments-in-thymeleaf -->
 <!-- https://www.baeldung.com/spring-thymeleaf-conditionals -->
 <!-- https://www.wimdeblauwe.com/blog/2021/09/14/thymeleaf-iteration-and-fragments/ -->
 <!-- https://www.appsloveworld.com/springboot/100/35/how-can-i-create-dynamic-table-in-thymeleaf -->
+
 <head th:fragment="head(title)">
 <meta charset="utf-8">
 <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -14,32 +16,25 @@
 <meta name="description" content="">
 <meta name="author" content="">
 <title th:replace="${title}">Planeador Académico</title>
-
 <!-- Custom fonts for this template-->
 <link th:href="@{/vendor/fontawesome-free/css/all.min.css}"
 	rel="stylesheet" type="text/css">
 <link
 	href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i"
 	rel="stylesheet">
-
 <!-- Custom styles for this template-->
 <link th:href="@{/css/sb-admin-2.min.css}" rel="stylesheet">
-
 <!-- Bootstrap core JavaScript-->
 <script th:src="@{/vendor/jquery/jquery.min.js}"></script>
 <script th:src="@{/vendor/bootstrap/js/bootstrap.bundle.min.js}"></script>
-
 <!-- Core plugin JavaScript-->
 <script th:src="@{/vendor/jquery-easing/jquery.easing.min.js}"></script>
-
 <!-- Custom scripts for all pages-->
 <script th:src="@{/js/sb-admin-2.min.js}"></script>
-
 <!-- Page level plugins -->
 <script th:src="@{/vendor/datatables/jquery.dataTables.min.js}"></script>
 <script th:src="@{/vendor/datatables/dataTables.bootstrap4.min.js}"></script>
-
 <!-- Page level custom scripts -->
-<script src="@{/js/demo/datatables-demo.js}"></script>
+<script th:src="@{/js/demo/datatables-demo.js}"></script>
 </head>
 </html>
