+++ src/main/resources/templates/fragments/sidebar.html
@@ -1,21 +1,16 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
 
-<!-- 		<li class="nav-item active"><a class="nav-link" -->
-<!-- 			th:href="@{/docente/logout}"> <i class="fas fa-fw fa-table"> </i> <span>Cerrar Sesión -->
-<!-- 			</span> -->
-<!-- 		</a></li> -->
-
-
+<!-- BARRA LATERAL PARA EL ADMINISTRADOR -->
 <ul th:fragment="_sidebar(user_type)" th:if="${user_type} == '_admin'"
 	class="navbar-nav bg-gradient-primary sidebar sidebar-dark accordion"
 	id="accordionSidebar">
 	<!-- Sidebar - Brand -->
 	<li class="nav-item"><a
 		class="sidebar-brand d-flex align-items-center justify-content-center"
-		href="index.html">
+		th:href="@{/admin/main}">
 			<div class="sidebar-brand-icon rotate-n-15">
-				<i class="far fa-calendar-alt" style="font-size:36px"> </i>
+				<i class="far fa-calendar-alt" style="font-size: 36px"></i>
 			</div>
 			<div class="sidebar-brand-text mx-3">
 				Sesión de <sup>Administrador</sup>
@@ -25,39 +20,25 @@
 	<!-- Nav Item - Dashboard -->
 	<li class="nav-item"><a class="nav-link"
 		th:href="@{/admin/microcurriculos/}"> <i
-			class="fas fa-fw fa-tachometer-alt"> </i> <span>Microcurrículos</span>
+			class="fas fa-fw fa-tachometer-alt"></i> <span>Microcurrículos</span>
 	</a></li>
 	<li class="nav-item"><a class="nav-link"
 		th:href="@{/admin/materias/}"> <i
-			class="fas fa-fw fa-tachometer-alt"> </i> <span>Materias</span>
+			class="fas fa-fw fa-tachometer-alt"></i> <span>Materias</span>
 	</a></li>
 	<!-- Divider -->
 	<hr class="sidebar-divider">
 	<!-- Heading -->
 	<div class="sidebar-heading">Dónde va esto</div>
-	<!-- Nav Item - Pages Collapse Menu -->
-	<!-- 		<li class="nav-item"><a class="nav-link collapsed" -->
-	<!-- 			th:href="@{/admin/solicitudes}" data-toggle="collapse" -->
-	<!-- 			data-target="#collapseTwo" aria-expanded="true" -->
-	<!-- 			aria-controls="collapseTwo"> <i class="fas fa-fw fa-cog"> </i> <span>Solicitudes</span> -->
-	<!-- 		</a> -->
-	<!-- 			<div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" -->
-	<!-- 				data-parent="#accordionSidebar"> -->
-	<!-- 				<div class="bg-white py-2 collapse-inner rounded"> -->
-	<!-- 					<h6 class="collapse-header">Custom Components:</h6> -->
-	<!-- 					<a class="collapse-item" href="buttons.html">Buttons </a> <a -->
-	<!-- 						class="collapse-item" href="cards.html">Cards </a> -->
-	<!-- 				</div> -->
-	<!-- 			</div></li> -->
 	<li class="nav-item"><a class="nav-link"
 		th:href="@{/admin/solicitudes}"> <i
-			class="fas fa-fw fa-tachometer-alt"> </i> <span>Solicitudes</span>
+			class="fas fa-fw fa-tachometer-alt"></i> <span>Solicitudes</span>
 	</a></li>
 	<!-- Nav Item - Utilities Collapse Menu -->
 	<li class="nav-item"><a class="nav-link collapsed" href="#"
 		data-toggle="collapse" data-target="#collapseUtilities"
 		aria-expanded="true" aria-controls="collapseUtilities"> <i
-			class="fas fa-fw fa-wrench"> </i> <span>Utilities </span>
+			class="fas fa-fw fa-wrench"></i> <span>Utilities </span>
 	</a>
 		<div id="collapseUtilities" class="collapse"
 			aria-labelledby="headingUtilities" data-parent="#accordionSidebar">
@@ -77,7 +58,7 @@
 	<li class="nav-item"><a class="nav-link collapsed" href="#"
 		data-toggle="collapse" data-target="#collapsePages"
 		aria-expanded="true" aria-controls="collapsePages"> <i
-			class="fas fa-fw fa-folder"> </i> <span>Pages </span>
+			class="fas fa-fw fa-folder"></i> <span>Pages </span>
 	</a>
 		<div id="collapsePages" class="collapse"
 			aria-labelledby="headingPages" data-parent="#accordionSidebar">
@@ -95,7 +76,7 @@
 		</div></li>
 	<!-- Nav Item - Charts -->
 	<li class="nav-item"><a class="nav-link" href="charts.html"> <i
-			class="fas fa-fw fa-chart-area"> </i> <span>Charts </span>
+			class="fas fa-fw fa-chart-area"></i> <span>Charts </span>
 	</a></li>
 	<!-- Divider -->
 	<hr class="sidebar-divider d-none d-md-block">
@@ -105,15 +86,21 @@
 	</div>
 </ul>
 
+<!-- SEPARADOR BARRA ORIGINAL -->
+<!-- SEPARADOR BARRA ORIGINAL -->
+<!-- SEPARADOR BARRA ORIGINAL -->
+<!-- SEPARADOR BARRA ORIGINAL -->
+
+<!-- BARRA LATERAL PARA EL DOCENTE -->
 <ul th:fragment="_sidebar(user_type)" th:if="${user_type} == '_docente'"
 	class="navbar-nav bg-gradient-primary sidebar sidebar-dark accordion"
 	id="accordionSidebar">
 	<!-- Sidebar - Brand -->
 	<li class="nav-item"><a
 		class="sidebar-brand d-flex align-items-center justify-content-center"
-		href="index.html">
+		th:href="@{/docente}">
 			<div class="sidebar-brand-icon rotate-n-15">
-				<i class="fas fa-laugh-wink"> </i>
+				<i class="fas fa-laugh-wink"></i>
 			</div>
 			<div class="sidebar-brand-text mx-3">
 				Sesión de <sup>Docente</sup>
@@ -123,7 +110,7 @@
 	<!-- Nav Item - Dashboard -->
 	<li class="nav-item"><a class="nav-link"
 		th:href="@{/admin/microcurriculos/}"> <i
-			class="fas fa-fw fa-tachometer-alt"> </i> <span>Planes</span>
+			class="fas fa-fw fa-tachometer-alt"></i> <span>Planes</span>
 	</a></li>
 	<!-- Divider -->
 	<hr class="sidebar-divider">
@@ -132,7 +119,7 @@
 	<!-- Nav Item - Pages Collapse Menu -->
 	<li class="nav-item"><a class="nav-link collapsed" href="#"
 		data-toggle="collapse" data-target="#collapseTwo" aria-expanded="true"
-		aria-controls="collapseTwo"> <i class="fas fa-fw fa-cog"> </i> <span>Cursos</span>
+		aria-controls="collapseTwo"> <i class="fas fa-fw fa-cog"></i> <span>Cursos</span>
 	</a>
 		<div id="collapseTwo" class="collapse" aria-labelledby="headingTwo"
 			data-parent="#accordionSidebar">
@@ -146,7 +133,7 @@
 	<li class="nav-item"><a class="nav-link collapsed" href="#"
 		data-toggle="collapse" data-target="#collapseUtilities"
 		aria-expanded="true" aria-controls="collapseUtilities"> <i
-			class="fas fa-fw fa-wrench"> </i> <span>Utilities </span>
+			class="fas fa-fw fa-wrench"></i> <span>Utilities </span>
 	</a>
 		<div id="collapseUtilities" class="collapse"
 			aria-labelledby="headingUtilities" data-parent="#accordionSidebar">
@@ -166,7 +153,7 @@
 	<li class="nav-item"><a class="nav-link collapsed" href="#"
 		data-toggle="collapse" data-target="#collapsePages"
 		aria-expanded="true" aria-controls="collapsePages"> <i
-			class="fas fa-fw fa-folder"> </i> <span>Pages </span>
+			class="fas fa-fw fa-folder"></i> <span>Pages </span>
 	</a>
 		<div id="collapsePages" class="collapse"
 			aria-labelledby="headingPages" data-parent="#accordionSidebar">
@@ -184,20 +171,13 @@
 		</div></li>
 	<!-- Nav Item - Charts -->
 	<li class="nav-item"><a class="nav-link" href="charts.html"> <i
-			class="fas fa-fw fa-chart-area"> </i> <span>Charts </span>
+			class="fas fa-fw fa-chart-area"></i> <span>Charts </span>
 	</a></li>
 	<!-- Divider -->
 	<hr class="sidebar-divider d-none d-md-block">
 	<!-- Sidebar Toggler (Sidebar) -->
 	<div class="text-center d-none d-md-inline">
-		<button class="rounded-circle border-0" id="sidebarToggle">
-
-
-		</button>
+		<button class="rounded-circle border-0" id="sidebarToggle"></button>
 	</div>
 </ul>
-
-
-
-
 </html>
