+++ src/main/resources/templates/main.html
@@ -1,44 +1,33 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-
+<html lang="en" xmlns:th="http://www.thymeleaf.org"
+	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
 <head th:replace="fragments/header :: head(~{::title})">
-	<title>Planeador Académico - Perfil de Administrador</title>
+<title>Planeador Académico - Perfil de Administrador</title>
 </head>
-
 <body id="page-top">
-
 	<!-- Page Wrapper -->
 	<div id="wrapper">
-
+		<!-- Sidebar -->
 		<ul
-			th:replace="${session.admin_id != 0} ? ~{fragments/sidebar :: _sidebar('_admin')} : ~{fragments/sidebar :: _sidebar('_docente')}">
-		</ul>
-
+			th:replace="${session.admin_id != 0} ? ~{fragments/sidebar :: _sidebar('_admin')} : ~{fragments/sidebar :: _sidebar('_docente')}"></ul>
 		<!-- Content Wrapper -->
 		<div id="content-wrapper" class="d-flex flex-column">
-
 			<!-- Main Content -->
 			<div id="content">
-
 				<!-- Topbar -->
 				<nav
-					th:replace="${session.admin_id != 0} ? ~{fragments/topbar :: top_bar('_admin')} : ~{fragments/topbar :: top_bar('_docente')}">
-				</nav>
-
-
+					th:replace="${session.admin_id != 0} ? ~{fragments/topbar :: top_bar('_admin')} : ~{fragments/topbar :: top_bar('_docente')}"></nav>
 				<!-- Begin Page Content -->
 				<div class="container-fluid">
-
 					<div class="row" style="text-align: center">
-
 						<div class="col-md-8">
 							<div class="card border-left-info shadow h-100 py-2">
 								<div class="card-body">
 									<div class="row no-gutters align-items-center">
 										<div class="col mr-2">
-											<div class="text-xs font-weight-bold text-info text-uppercase mb-1">
-												Planeadores terminados
-											</div>
+											<div
+												class="text-xs font-weight-bold text-info text-uppercase mb-1">
+												Planeadores terminados</div>
 											<div class="row no-gutters align-items-center">
 												<div class="col-auto">
 													<div class="h5 mb-0 mr-3 font-weight-bold text-gray-800">90%</div>
@@ -58,16 +47,14 @@
 									</div>
 								</div>
 							</div>
-
 						</div>
-
-
 						<div class="col-md-4">
 							<div class="card border-left-warning shadow h-100 py-2">
 								<div class="card-body">
 									<div class="row no-gutters align-items-center">
 										<div class="col mr-2">
-											<div class="text-xs font-weight-bold text-warning text-uppercase mb-1">
+											<div
+												class="text-xs font-weight-bold text-warning text-uppercase mb-1">
 												Solicitudes pendientes</div>
 											<div class="h5 mb-0 font-weight-bold text-gray-800">18</div>
 										</div>
@@ -78,17 +65,9 @@
 								</div>
 							</div>
 						</div>
-
 					</div>
-
-					<div class="row">
-
-						&nbsp;
-
-					</div>
+					<div class="row">&nbsp;</div>
 					<div class="row" style="text-align: center;">
-
-
 						<div class="col-md-8">
 							<div class="card shadow mb-4">
 								<div class="card-header py-3">
@@ -96,7 +75,8 @@
 								</div>
 								<div class="card-body">
 									<div class="table-responsive">
-										<table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
+										<table class="table table-bordered" id="dataTable"
+											width="100%" cellspacing="0">
 											<thead>
 												<tr>
 													<th>Nombre</th>
@@ -117,37 +97,31 @@
 													<td th:text="${adm.email}"></td>
 													<td th:text="${adm.nombre}"></td>
 													<td th:text="${adm.nombre}"></td>
-
 												</tr>
 												<tr>
 													<td th:text="${adm.nombre}"></td>
 													<td th:text="${adm.email}"></td>
 													<td th:text="${adm.nombre}"></td>
 													<td th:text="${adm.nombre}"></td>
-
 												</tr>
 												<tr>
 													<td th:text="${adm.nombre}"></td>
 													<td th:text="${adm.email}"></td>
 													<td th:text="${adm.nombre}"></td>
 													<td th:text="${adm.nombre}"></td>
-
 												</tr>
 												<tr>
 													<td th:text="${adm.nombre}"></td>
 													<td th:text="${adm.email}"></td>
 													<td th:text="${adm.nombre}"></td>
 													<td th:text="${adm.nombre}"></td>
-
 												</tr>
 											</tbody>
 										</table>
 									</div>
 								</div>
 							</div>
-
 						</div>
-
 						<div class="col-md-4">
 							<div class="card shadow mb-4">
 								<div class="card-header py-3">
@@ -155,11 +129,11 @@
 								</div>
 								<div class="card-body">
 									<div class="table-responsive">
-										<table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
+										<table class="table table-bordered" id="dataTable"
+											width="100%" cellspacing="0">
 											<thead>
 												<tr>
 													<th>Nombre</th>
-
 												</tr>
 											</thead>
 											<tbody th:each="adm: ${admin}">
@@ -169,86 +143,36 @@
 												<tr>
 													<td th:text="${adm.nombre}"></td>
 												</tr>
-										
 												<tr>
 													<td th:text="${adm.nombre}"></td>
 												</tr>
-										
 												<tr>
 													<td th:text="${adm.nombre}"></td>
 												</tr>
-										
 												<tr>
 													<td th:text="${adm.nombre}"></td>
 												</tr>
-										
-										
-
 											</tbody>
 										</table>
 									</div>
 								</div>
 							</div>
-
 						</div>
-
 					</div>
-
 				</div>
 				<!-- /.container-fluid -->
 			</div>
-
-
-
+			<!-- End of Main Content -->
 		</div>
-		<!-- End of Main Content -->
-
-
-
-
-
-
-	</div>
-	<!-- End of Content Wrapper -->
-
+		<!-- End of Content Wrapper -->
 	</div>
 	<!-- End of Page Wrapper -->
-
 	<!-- Scroll to Top Button-->
-	<a class="scroll-to-top rounded" href="#page-top"> <i class="fas fa-angle-up"></i>
+	<a class="scroll-to-top rounded" href="#page-top"> <i
+		class="fas fa-angle-up"></i>
 	</a>
-
+	<!-- Logout Modal-->
 	<div
-		th:replace="${session.admin_id != 0} ? ~{fragments/topbar :: logout_modal('_admin')} : ~{fragments/topbar :: logout_modal('_docente')}">
-	</div>
-
-
-
-
-	<!-- 	<div class="modal fade" id="logoutModal" tabindex="-1" role="dialog" -->
-	<!-- 		aria-labelledby="exampleModalLabel" aria-hidden="true"> -->
-	<!-- 		<div class="modal-dialog" role="document"> -->
-	<!-- 			<div class="modal-content"> -->
-	<!-- 				<div class="modal-header"> -->
-	<!-- 					<h5 class="modal-title" id="exampleModalLabel">Abandono de -->
-	<!-- 						Sesión</h5> -->
-	<!-- 					<button class="close" type="button" data-dismiss="modal" -->
-	<!-- 						aria-label="Close"> -->
-	<!-- 						<span aria-hidden="true">×</span> -->
-	<!-- 					</button> -->
-	<!-- 				</div> -->
-	<!-- 				<div class="modal-body">Confirme que desea salir de la -->
-	<!-- 					aplicación.</div> -->
-	<!-- 				<div class="modal-footer"> -->
-	<!-- 									Posiblemente borrar este botón en el futuro -->
-	<!-- 					<button class="btn btn-secondary" type="button" -->
-	<!-- 						data-dismiss="modal">Cancelar</button> -->
-	<!-- 					<a class="btn btn-primary" th:href="@{/admin/logout}">Cerrar -->
-	<!-- 						Sesión</a> -->
-	<!-- 				</div> -->
-	<!-- 			</div> -->
-	<!-- 		</div> -->
-	<!-- 	</div> -->
+		th:replace="${session.admin_id != 0} ? ~{fragments/topbar :: logout_modal('_admin')} : ~{fragments/topbar :: logout_modal('_docente')}"></div>
 </body>
-
 </html>
