+++ src/main/resources/templates/register_admin.html
