+++ src/main/resources/templates/solicitudes.html
@@ -1,8 +1,9 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+<html lang="en" xmlns:th="http://www.thymeleaf.org"
+	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
 
 <head th:replace="fragments/header :: head(~{::title})">
-	<title>Planeador Académico - Lista de Microcurrículos</title>
+<title>Planeador Académico - Lista de Microcurrículos</title>
 </head>
 
 <body id="page-top">
@@ -26,17 +27,22 @@
 					th:replace="${session.admin_id != null} ? ~{fragments/topbar :: top_bar(_admin)} : ~{fragments/topbar :: top_bar(_docente)}">
 				</nav>
 
-				<div th:if="${msjAprobacion != null}" class="alert alert-primary alert-dismissible fade show"
+				<div th:if="${msjAprobacion != null}"
+					class="alert alert-primary alert-dismissible fade show"
 					role="alert">
 					<label th:text="${msjRechazo}"></label>
-					<button type="button" class="close" data-dismiss="alert" aria-label="Close">
+					<button type="button" class="close" data-dismiss="alert"
+						aria-label="Close">
 						<span aria-hidden="true">&times;</span>
 					</button>
 				</div>
 
-				<div th:if="${msjRechazo != null}" class="alert alert-primary alert-dismissible fade show" role="alert">
+				<div th:if="${msjRechazo != null}"
+					class="alert alert-primary alert-dismissible fade show"
+					role="alert">
 					<label th:text="${msjRechazo}"></label>
-					<button type="button" class="close" data-dismiss="alert" aria-label="Close">
+					<button type="button" class="close" data-dismiss="alert"
+						aria-label="Close">
 						<span aria-hidden="true">&times;</span>
 					</button>
 				</div>
@@ -46,19 +52,22 @@
 
 					<div class="row">
 
-						<div class="col-md-12"  style="text-align: center;">
+						<div class="col-md-12" style="text-align: center;">
 
 
 
 							<div class="card shadow mb-4">
 								<div class="card-header py-3">
-									<h6 class="m-0 font-weight-bold text-primary">Solicitudes de
-										admisión</h6>
+									<h6 class="m-0 font-weight-bold text-primary">Solicitudes
+										de admisión</h6>
 								</div>
 
 								<div class="card-body">
 									<div class="table-responsive">
-										<table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
+										<!-- Utiliza un Map y una lista de campos para dibujar todos los campos de la tabla solicitada -->
+										<!-- Referirse a AdministradorControlador para más detalles. -->
+										<table class="table table-bordered" id="dataTable"
+											width="100%" cellspacing="0">
 											<thead>
 
 												<tr>
@@ -67,16 +76,20 @@
 											</thead>
 											<tbody>
 												<tr th:each="row: ${rows}">
-													<td th:each="header: ${headers}"><a class="btn btn-primary"
-															th:if="${header == 'Aprobación'}"
-															th:href="@{/admin/aprobar/{id}(id=${row.get('Aprobación')})}">Aprobar</a>
+													<td th:each="header: ${headers}">
+														<!-- Evalúa condicionalmente el valor de uno de los headers para dibujar una celda de forma diferente -->
+														<!-- En este caso, para crear botones interactivos con las acciones de aprobar o rechazar solicitudes de registro del docente -->
+														<a class="btn btn-primary"
+														th:if="${header == 'Aprobación'}"
+														th:href="@{/admin/aprobar/{id}(id=${row.get('Aprobación')})}">Aprobar</a>
 
-														<a class="btn btn-danger" th:if="${header == 'Aprobación'}"
-															th:href="@{/admin/rechazar/{id}(id=${row.get('Aprobación')})}"
-															onclick="return confirm('Confirme que desea proceder a rechazar esta solicitud');">Rechazar</a>
+														<a class="btn btn-danger"
+														th:if="${header == 'Aprobación'}"
+														th:href="@{/admin/rechazar/{id}(id=${row.get('Aprobación')})}"
+														onclick="return confirm('Confirme que desea proceder a rechazar esta solicitud');">Rechazar</a>
 
 														<span th:if="${header != 'Aprobación'}"
-															th:text="${row.get(header)}"></span>
+														th:text="${row.get(header)}"></span>
 													</td>
 												</tr>
 											</tbody>
@@ -84,24 +97,22 @@
 									</div>
 								</div>
 							</div>
-
-
-
 						</div>
-
 					</div>
 
-
 				</div>
+			</div>
+		</div>
+	</div>
 
+	<!-- Scroll to Top Button-->
+	<a class="scroll-to-top rounded" href="#page-top"> <i
+		class="fas fa-angle-up"></i>
+	</a>
 
-				<!-- Scroll to Top Button-->
-				<a class="scroll-to-top rounded" href="#page-top"> <i class="fas fa-angle-up"></i>
-				</a>
-
-				<div
-					th:replace="${session.admin_id != 0} ? ~{fragments/topbar :: logout_modal('_admin')} : ~{fragments/topbar :: logout_modal('_docente')}">
-				</div>
+	<div
+		th:replace="${session.admin_id != 0} ? ~{fragments/topbar :: logout_modal('_admin')} : ~{fragments/topbar :: logout_modal('_docente')}">
+	</div>
 </body>
 
 </html>
